Delete other clear Commands, implement only 1 clear-all command
